Ultroid v0.7 Updates
Date: 30-8-2022

+++ plugins/pmpermit.py
@@ -200,8 +200,11 @@ if udB.get_key("PMSETTING"):
             incoming=True,
             func=lambda e: e.is_private
             and e.sender_id not in DEVLIST
-            and not (e.out or e.sender.bot or e.sender.is_self or e.sender.verified),
-        ),
+            and not e.out
+            and not e.sender.bot
+            and not e.sender.is_self
+            and not e.sender.verified,
+        )
     )
     async def permitpm(event):
         inline_pm = Redis("INLINE_PM") or False
@@ -212,7 +215,7 @@ if udB.get_key("PMSETTING"):
                     await ultroid_bot.edit_folder(user.id, folder=1)
                 except BaseException as er:
                     LOGS.info(er)
-            if event.media:
+            if event.media and not udB.get_key("DISABLE_PMDEL"):
                 await event.delete()
             name = user.first_name
             fullname = get_display_name(user)
@@ -507,11 +510,7 @@ async def blockpm(block):
         user = block.chat_id
     else:
         return await eor(block, NO_REPLY, time=10)
-    if user in DEVLIST:
-        return await eor(
-            block,
-            "`Lol, He is my Developer\nHe Can't Be Blocked`",
-        )
+
     await block.client(BlockRequest(user))
     aname = await block.client.get_entity(user)
     await block.eor(f"{inline_mention(aname)} [`{user}`] `has been blocked!`")
@@ -621,14 +620,13 @@ async def list_approved(event):
                 tabulate(users, headers=["UserName", "UserID"], showindex="always")
             )
         else:
-            text = ""
-            for user in users:
-                text += f"[{user[-1]}] - {user[0]}"
+            text = "".join(f"[{user[-1]}] - {user[0]}" for user in users)
             list_appr.write(text)
     await event.reply(
-        "List of users approved by [{}](tg://user?id={})".format(OWNER_NAME, OWNER_ID),
+        f"List of users approved by [{OWNER_NAME}](tg://user?id={OWNER_ID})",
         file="approved_pms.txt",
     )
+
     await xx.delete()
     remove("approved_pms.txt")
 
@@ -813,7 +811,7 @@ async def in_pm_ans(event):
                 await event.builder.document(
                     res,
                     title="Inline PmPermit",
-                    description="~ @TheUltroid",
+                    description="~ @TeamUltroid",
                     text=msg_,
                     buttons=buttons,
                     link_preview=False,
